refactor : delete Players
 - 불필요한 일급컬렉션 삭제
