+++ src/main/java/view/InputView.java
@@ -1,5 +1,7 @@
 package view;
 
+import view.dto.PlayerDto;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
