+++ src/main/java/view/OutputView.java
@@ -2,9 +2,9 @@ package view;
 
 import domain.user.Dealer;
 import domain.user.Gamer;
-import domain.user.Players;
+import domain.user.Gamers;
+import view.dto.CardStatus;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class OutputView {
@@ -12,20 +12,11 @@ public class OutputView {
     private static final String JOIN_DELIMITER = ", ";
     private static final String CARD_STATUS_FORMAT = "%s 카드 : %s";
 
-    public static void showStartStatus(Dealer dealer, Players players) {
-        String names = String.join(JOIN_DELIMITER, players.getNames());
+    public static void showStartStatus(Gamers gamers) {
+        String names = String.join(JOIN_DELIMITER, gamers.getPlayerName());
         System.out.println("딜러와 " + names + "에게 2장의 카드를 나누었습니다.\n");
 
-        List<Gamer> gamers = collectGamer(dealer, players);
-
-        showAllGamerStatus(gamers);
-    }
-
-    private static List<Gamer> collectGamer(Dealer dealer, Players players) {
-        List<Gamer> gamers = new ArrayList<>();
-        gamers.add(dealer);
-        gamers.addAll(players.getPlayers());
-        return gamers;
+        showAllGamerStatus(gamers.getGamers());
     }
 
     private static void showAllGamerStatus(List<Gamer> gamers) {
@@ -47,10 +38,10 @@ public class OutputView {
 
     public static void showDealerCanReceive(boolean canReceive) {
         if (canReceive) {
-            System.out.println("딜러는 16이하라 한장의 카드를 더 받습니다.\n");
+            System.out.println("\n딜러는 16이하라 한장의 카드를 더 받습니다.\n");
             return;
         }
-        System.out.println("딜러는 17이상이라 카드를 받지 않습니다.\n");
+        System.out.println("\n딜러는 17이상이라 카드를 받지 않습니다.\n");
     }
 
     public static void showLastStatus(List<Gamer> gamers) {
@@ -61,7 +52,8 @@ public class OutputView {
         }
     }
 
-    public static void showResult() {
+    public static void showResult(Dealer dealer, List<Gamer> players) {
         System.out.println("## 최종 수익");
+
     }
 }
