+++ src/main/java/view/dto/CardStatus.java
@@ -1,4 +1,4 @@
-package view;
+package view.dto;
 
 import domain.card.Card;
