+++ src/main/java/view/dto/PlayerDto.java
@@ -1,4 +1,4 @@
-package view;
+package view.dto;
 
 import domain.user.Player;
